fix: add missing jsfiddle.sh check for react samples

+++ samples/jsfiddle.sh
@@ -27,10 +27,10 @@ echo "NAME: ${NAME}"
 mkdir -p "${DIST_DIR}/samples/${NAME}/jsfiddle"
 
 # Copy files
-echo "Copy ${SCRIPT_DIR}/${NAME}/index.js to ${DIST_DIR}/samples/${NAME}/jsfiddle/demo.js"
-cp "${SCRIPT_DIR}/${NAME}/index.js" "${DIST_DIR}/samples/${NAME}/jsfiddle/demo.js"
-cp "${SCRIPT_DIR}/${NAME}/index.html" "${DIST_DIR}/samples/${NAME}/jsfiddle/demo.html"
-cp "${SCRIPT_DIR}/${NAME}/style.css" "${DIST_DIR}/samples/${NAME}/jsfiddle/demo.css"
+echo "Copy files to ${DIST_DIR}/samples/${NAME}/jsfiddle/"
+[ -f "${SCRIPT_DIR}/${NAME}/index.js" ] && cp "${SCRIPT_DIR}/${NAME}/index.js" "${DIST_DIR}/samples/${NAME}/jsfiddle/demo.js"
+[ -f "${SCRIPT_DIR}/${NAME}/index.html" ] && cp "${SCRIPT_DIR}/${NAME}/index.html" "${DIST_DIR}/samples/${NAME}/jsfiddle/demo.html"
+[ -f "${SCRIPT_DIR}/${NAME}/style.css" ] && cp "${SCRIPT_DIR}/${NAME}/style.css" "${DIST_DIR}/samples/${NAME}/jsfiddle/demo.css"
 
 # Copy the public folder if one is found (graphics, other static files).
 if [ -d "public" ] && [ "$(ls -A public)" ]; then
@@ -46,21 +46,25 @@ pushd "${DIST_DIR}/samples/${NAME}/jsfiddle"
 
 # Remove region tags from files by type, since they all have different comment conventions.
 # We use a conditional here since sed behaves differently on Linux.
-echo "Remove region tags from ${DIST_DIR}/samples/${NAME}/jsfiddle/demo.js"
-sed -i.sed-back "s#// \[START .*]##g" "demo.js" && rm *.sed-back
-sed -i.sed-back "s#// \[END .*]##g" "demo.js" && rm *.sed-back
-sed -i.sed-back "s#/\* \[START .*] \*/##g" "demo.js" && rm *.sed-back
-sed -i.sed-back "s#/\* \[END .*] \*/##g" "demo.js"  && rm *.sed-back
+echo "Remove region tags from JS"
+if [ -f "demo.js" ]; then
+  sed -i.sed-back "s#// \[START .*]##g" "demo.js" && rm *.sed-back
+  sed -i.sed-back "s#// \[END .*]##g" "demo.js" && rm *.sed-back
+  sed -i.sed-back "s#/\* \[START .*] \*/##g" "demo.js" && rm *.sed-back
+  sed -i.sed-back "s#/\* \[END .*] \*/##g" "demo.js"  && rm *.sed-back
+fi
 
+echo "Remove region tags from HTML"
+if [ -f "demo.html" ]; then
+  sed -i.sed-back "s/<!--[[:space:]]*\[START .*\][[:space:]]*-->//g" "demo.html" && rm *.sed-back
+  sed -i.sed-back "s/<!--[[:space:]]*\[END .*\][[:space:]]*-->//g" "demo.html" && rm *.sed-back
+fi
 
-echo "Remove region tags from ${DIST_DIR}/samples/${NAME}/jsfiddle/demo.html"
-sed -i.sed-back "s/<!--[[:space:]]*\[START .*\][[:space:]]*-->//g" "demo.html" && rm *.sed-back
-sed -i.sed-back "s/<!--[[:space:]]*\[END .*\][[:space:]]*-->//g" "demo.html" && rm *.sed-back
-
-
-echo "Remove region tags from ${DIST_DIR}/samples/${NAME}/jsfiddle/demo.css"
-sed -i.sed-back "s#/\* \[START maps_.*] \*/##g" "demo.css" && rm *.sed-back
-sed -i.sed-back "s#/\* \[END maps_.*] \*/##g" "demo.css"  && rm *.sed-back
+echo "Remove region tags from CSS"
+if [ -f "demo.css" ]; then
+  sed -i.sed-back "s#/\* \[START maps_.*] \*/##g" "demo.css" && rm *.sed-back
+  sed -i.sed-back "s#/\* \[END maps_.*] \*/##g" "demo.css"  && rm *.sed-back
+fi
 
 # after that other cleanup, re-run prettier to remove excess newlines
 npx prettier -w . --ignore-path /dev/null
